Progress on night actions
