+++ MattEland.WhereDoggo/MattEland.WhereDoggo.Core/Roles/ApprenticeSeerNightAction.cs
@@ -8,10 +8,13 @@ public class ApprenticeSeerNightAction : RoleNightActionBase
     {
     }
 
+    /// <inheritdoc />
     public override decimal NightActionOrder => 5.1m;
 
+    /// <inheritdoc />
     public override string WakeInstructions => "Apprentice Seer, choose a card in the middle to view";
 
+    /// <inheritdoc />
     public override void PerformNightAction(Game game, GamePlayer player)
     {
         IHasCard? slot = player.PickSingleCard(game.CenterSlots.Cast<IHasCard>().ToList());
