+++ MattEland.WhereDoggo/MattEland.WhereDoggo.Core/Roles/MysticWolfRole.cs
@@ -1,4 +1,6 @@
-﻿namespace MattEland.WhereDoggo.Core.Roles;
+﻿using MattEland.WhereDoggo.Core.Engine.Phases;
+
+namespace MattEland.WhereDoggo.Core.Roles;
 
 /// <summary>
 /// The Mystic Wolf is a werewolf that may look at one player's card during the night 
@@ -13,28 +15,15 @@ public class MysticWolfRole : WerewolfRole
     /// <inheritdoc />
     public override Teams Team => Teams.Werewolves;
 
-    /// <inheritdoc />
-    public override decimal? NightActionOrder => 2.2m;
-    
 
     /// <inheritdoc />
-    public override void PerformNightAction(Game game, GamePlayer player)
+    public override IEnumerable<NightActionBase> NightActions
     {
-        base.PerformNightAction(game, player);
-
-        IEnumerable<IHasCard> otherPlayerTargets = game.GetOtherPlayerTargets(player);
-        IDictionary<IHasCard, CardProbabilities> probs = player.Brain.BuildInitialRoleProbabilities();
-        IHasCard? cardHolder = player.PickSingleCard(otherPlayerTargets.Where(t => probs[t].CalculateTeamProbability(Teams.Werewolves) < 1m));
-
-        if (cardHolder == null)
+        get
         {
-            game.LogEvent(new SkippedNightActionEvent(player));
-        }
-        else
-        {
-            game.LogEvent(new ObservedPlayerCardEvent(player, cardHolder));
+            yield return new WerewolfNightAction();
+            yield return new MysticWolfNightAction();
         }
     }
-    
-    
+
 }
