+++ MattEland.WhereDoggo/MattEland.WhereDoggo.Core/Roles/RevealerRole.cs
@@ -1,4 +1,6 @@
-﻿namespace MattEland.WhereDoggo.Core.Roles;
+﻿using MattEland.WhereDoggo.Core.Engine.Phases;
+
+namespace MattEland.WhereDoggo.Core.Roles;
 
 /// <summary>
 /// The Revealer from One Night Ultimate Werewolf Daybreak
@@ -16,31 +18,11 @@ public class RevealerRole : CardBase
     public override Teams Team => Teams.Villagers;
 
     /// <inheritdoc />
-    public override decimal? NightActionOrder => 10m;
-    
-    /// <inheritdoc />
-    public override void PerformNightAction(Game game, GamePlayer player)
+    public override IEnumerable<NightActionBase> NightActions
     {
-        IHasCard? target = player.PickSingleCard(game.GetOtherPlayerTargets(player));
-
-        if (target == null)
+        get
         {
-            game.LogEvent(new SkippedNightActionEvent(player));
-        }
-        else
-        {
-            GamePlayer targetPlayer = (GamePlayer)target;
-
-            targetPlayer.CurrentCard.IsRevealed = true;
-            game.LogEvent(new RevealedRoleEvent(player, targetPlayer));
-            game.LogEvent(new RevealedRoleObservedEvent(player, targetPlayer));
-
-            // Only villagers should be revealed
-            if (targetPlayer.CurrentCard.Team != Teams.Villagers)
-            {
-                targetPlayer.CurrentCard.IsRevealed = false;
-                game.LogEvent(new RevealerHidEvilRoleEvent(player, targetPlayer));
-            }
+            yield return new RevealerNightAction();
         }
     }
 }
