+++ MattEland.WhereDoggo/MattEland.WhereDoggo.Core/Roles/SeerRole.cs
@@ -1,4 +1,6 @@
-﻿namespace MattEland.WhereDoggo.Core.Roles;
+﻿using MattEland.WhereDoggo.Core.Engine.Phases;
+
+namespace MattEland.WhereDoggo.Core.Roles;
 
 /// <summary>
 /// The seer from One Night Ultimate Werewolf.
@@ -14,33 +16,13 @@ public class SeerRole : CardBase
     /// <inheritdoc />
     public override Teams Team => Teams.Villagers;
 
-    /// <inheritdoc />
-    public override decimal? NightActionOrder => 5m;
 
     /// <inheritdoc />
-    public override void PerformNightAction(Game game, GamePlayer player)
+    public override IEnumerable<NightActionBase> NightActions
     {
-        base.PerformNightAction(game, player);
-        
-        // Choose whether we're skipping, getting another player, or getting 2 center cards
-        IList<IHasCard> playerChoices = game.Players.Where(p => p != player).Cast<IHasCard>().ToList(); 
-        IList<IHasCard> centerChoices = game.CenterSlots.Cast<IHasCard>().ToList();
-        List<IHasCard> cards = player.PickSeerCards(playerChoices, centerChoices);
-
-        switch (cards.Count)
+        get
         {
-            case <= 0:
-                // A seer should really never skip their night action. You're just losing information.
-                game.LogEvent(new SkippedNightActionEvent(player));
-                break;
-            case 1:
-                // Look at the other player's card
-                game.LogEvent(new ObservedPlayerCardEvent(player, cards.Single()));
-                break;
-            default:
-                // Look at the center cards
-                cards.ForEach(c => game.LogEvent(new ObservedCenterCardEvent(player, c)));
-                break;
+            yield return new SeerNightAction();
         }
     }
 }
