+++ MattEland.WhereDoggo/MattEland.WhereDoggo.Core/Roles/SentinelRole.cs
@@ -1,4 +1,6 @@
-﻿namespace MattEland.WhereDoggo.Core.Roles;
+﻿using MattEland.WhereDoggo.Core.Engine.Phases;
+
+namespace MattEland.WhereDoggo.Core.Roles;
 
 /// <summary>
 /// The sentinel from One Night Ultimate Werewolf Daybreak
@@ -15,22 +17,11 @@ public class SentinelRole : CardBase
     public override RoleTypes RoleType => RoleTypes.Sentinel;
 
     /// <inheritdoc />
-    public override decimal? NightActionOrder => 0.1m;
-
-    /// <inheritdoc />
-    public override void PerformNightAction(Game game, GamePlayer player)
+    public override IEnumerable<NightActionBase> NightActions
     {
-        // Sentinels may choose to skip placing their token
-        List<IHasCard> otherPlayers = game.Players.Where(p => p != player).Cast<IHasCard>().ToList();
-        if (player.PickSingleCard(otherPlayers) is GamePlayer target)
+        get
         {
-            target.HasSentinelToken = true;
-            game.LogEvent(new SentinelTokenPlacedEvent(player, target));
-            game.LogEvent(new SentinelTokenObservedEvent(player, target));
-        }
-        else
-        {
-            game.LogEvent(new SkippedNightActionEvent(player));
+            yield return new SentinelNightAction();
         }
     }
 }
