+++ MattEland.WhereDoggo/MattEland.WhereDoggo.WPFClient/Controls/GamePage.xaml
@@ -87,7 +87,7 @@
 
                         <!-- Available Roles -->
                         <StackPanel Orientation="Vertical" DockPanel.Dock="Right">
-                            <Label Content="In-Game Role Order" FontWeight="Bold" Target="{Binding ElementName=AvailableRoles}" />
+                            <Label Content="In-Game Roles" FontWeight="Bold" Target="{Binding ElementName=AvailableRoles}" />
                             <ItemsControl ItemsSource="{Binding Path=Roles}" x:Name="AvailableRoles" HorizontalContentAlignment="Left" VerticalContentAlignment="Center" Padding="8" HorizontalAlignment="Left">
                                 <ItemsControl.ItemsPanel>
                                     <ItemsPanelTemplate>
